Function #1, with size_t type for the number of nodes

+++ 0x12-more_singly_linked_lists/0-print_listint.c
@@ -11,7 +11,7 @@
 size_t print_listint(const listint_t *h)
 {
 	listint_t *current = (listint_t *)h;
-	int nodes = 0;
+	size_t nodes = 0;
 
 	while (current)
 	{
